Moves to compose-layout
Adds as default to AppScaffold
