+++ compose-layout/src/test/snapshots/images/com.google.android.horologist.compose.layout_ResponsiveTimeTextTest_defaultTimeText.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d0acfdca65c8f367c190285c06c185af0b33477e439d1133ec67067862b44eeb
+size 13194
